api/requirements.txt
fastapi
uvicorn[standard]
qdrant-client
pandas
pydantic
sentence-transformers
scikit-learn
requests
python-dotenv
numpy<2.0
